servidor/requirements.txt
click==7.1.2
Flask==0.12.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
Werkzeug==1.0.1
